Move Comment to core model
