+++ gaphor/core/modeling/__init__.py
@@ -1,4 +1,9 @@
-from gaphor.core.modeling.coremodel import Element, NamedElement, PackageableElement
+from gaphor.core.modeling.coremodel import (
+    Comment,
+    Element,
+    NamedElement,
+    PackageableElement,
+)
 from gaphor.core.modeling.diagram import Diagram
 from gaphor.core.modeling.elementfactory import ElementFactory
 from gaphor.core.modeling.event import *
